Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -16,11 +16,11 @@ int main(void)
 	char d;
 	float f;
 
-	printf("Size of a char: %d byte(s)\n", (unsigned long) sizeof(d));
-	printf("Size of an int: %d byte(s)\n", (unsigned long) sizeof(a));
-	printf("Size of a long int: %d byte((s)\n", (unsigned long) sizeof(b));
-	printf("Size of a long long int: %d byte(s)\n", (unsigned long) sizeof(c));
-	printf("Size of a float: %d byte(s)\n", (unsigned long) sizeof(f));
+	printf("Size of a char: %d byte(s)\n", (int) sizeof(d));
+	printf("Size of an int: %d byte(s)\n", (int) sizeof(a));
+	printf("Size of a long int: %d byte((s)\n", (int) sizeof(b));
+	printf("Size of a long long int: %d byte(s)\n", (int) sizeof(c));
+	printf("Size of a float: %d byte(s)\n", (int) sizeof(f));
 
 	return (0);
 }
